PageResponse total type from int to int64

+++ http/response/paginate.go
@@ -1,11 +1,11 @@
 package response
 
 type PaginateResponse[T any] struct {
-	Total int `json:"total"`
-	Data  []T `json:"data"`
+	Total int64 `json:"total"`
+	Data  []T   `json:"data"`
 }
 
-func NewPaginateResponse[T any](total int, data []T) *PaginateResponse[T] {
+func NewPaginateResponse[T any](total int64, data []T) *PaginateResponse[T] {
 	return &PaginateResponse[T]{
 		Total: total,
 		Data:  data,
